pyproject.toml
[project]
name = "youqu3"
description = "YouQu-Next"
authors = [
    { name = "mikigo", email = "huangmingqiang@uniontech.com" },
]
dependencies = [
    "pytest-timeout",
    "pytest-rerunfailures",
    "pytest-mark-manage",
    "allure-pytest",
    "pytest-json-report",
    "funnylog2",
    "click",
    "pexpect",
]
dynamic = ["version"]

requires-python = ">=3.7"
readme = "README.md"
license = { text = "GPL2.0" }

[project.optional-dependencies]
gui = [
    "pylinuxauto",
    "youqu-html-rpc",
    "pytest-record-video",
]
webui = [
    "playwright",
    "youqu-html-rpc",
    "pytest-record-video",
]
http = [
    "requests",
    "youqu-html-rpc",
]
remote = [
    "fabric",
]
local-html = [
    "youqu-html",
]

[project.scripts]
youqu3 = "youqu3.driver.cli:cli"

[project.entry-points.pytest11]
pytest-youqu-plugin = "youqu3.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch]
version = { path = "youqu3/__version__.py" }

[tool.hatch.build.targets.wheel]
packages = [
    "/youqu3",
]

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/.venv",
  "/.env",
  "/.gitignore",
  "/dist",
  "/docs",
  "/node_modules",
  "/tests",
  "/package.json",
  "/Pipfile",
  "/publish.sh",
]
